Update dialog_quickmessage.xml
first thank you very much for sharing such kind of works. I found a delete button in quick message popup in a nice messaging app Please update this. I thinks this would be nice if we get a delete button in quick message popup, as sometimes we get sms from service provider or from somewhere which is not important; we can delete it from the popup itself rather going to the main messaging app.
thanks